Re #118 - applied workaround for malformed nxspe files
In old mantid versions, and in some corner cases with newer Mantid,
SaveNXSPE does not save the instrument name in the NXSPE file. Thus
LoadNXSPE does not know if the data is from a Direct or Indirect
geometry instrument and sets EMode to Elastic. This causes problem
in calculating the projection and limits. The NXSPE files usually
define a log entry 'Ei' which is set to NaN for indirect. The
workaround checks if this entry exists and sets EMode base on it.