Replace quick options presenters re #294
A lot of the code used to create the various presenter classes was repeated
and would have to be duplicated if more options were added. This adds a
higher-order function run_quick_options instead.

Also get rid of the quick_options_presenter attribute in the plot classes as
it isn't used.