Fix other intensity corrections re #358
Since all the plotting is now done with workspaces and the SlicheCache
class, some refactoring was needed for the intensity methods. Part of
this involved fixing the way binary operations work on mslice
workspaces e.g. division.

Whether the data is rotated or not is dealt with at a higher level so
the compute_ methods no longer need this parameter.