Fix bug with quick options for slice overlay lines
Sharing the quick options code between slices and cuts means
coping with some options being unavilable, e.g. error bars
for recoil lines overlayed on slices. This now checks before
attempting to create options that don't apply to the plot
type.
Also fixes an issue with mapping colors to their names and back.
The logic is moved to a separate testable module and uses the
default cycler behaviour rather than duplicating this in mslice
to be more robust against changes in the future.