Allow selection of more than one workspace for addition in workspace panel in MSlice
Fixed bug that only one workspace can be selected from the workspace panel so the Add Workspace dialog had to be used to select additional workspaces. Now the Add Workspace dialog only opens when one workspace was selected from the workspace panel.
In case of a selection of more than one workspace clicking the Add button is enough to add the selected workspaces.