Use deletion callback to remove the line from the plot options
Refs #776. This fixes a crash caused by a delay in the 'destroyed'
signal being emitted for a LegendAndLineOptionsSetter. I believe this
delay is due to the deleteLater slot being triggered whilst the event
loop is not running. This means the destroy signal is only called when
the event loop is restarted, and at this point the plot options were
already closed, and so there was a crash.