Added opiton to scale fonts
Added an option to scale all fonts in a slice and cut plot as to preserve the proportion of the font sizes.
Changed the getter and setter of plot title as to change only the text in the title, and not other parameters like font size.
All functunality is working except when plot options window is set to open.
When window is not closed, strange things happen when using the scaling option.