Added all missing functunality and corrected bugs
Devised a possibly over-engineered solution to provide the best experience for the users.
This implementation uses two checkboxes and two text fieds to control setting and scaling of fonts.
Tested manually on slice and cut plots, with closing and keeping the options window open.
The GUI prevents users from selecting scaling and setting simultaneously.

Short explanation:

When any parameters are changed in the GUI of plot options, they are appended to temporary dictionary.
When the OK button is pressed, all changes stored in the temp dict are loaded onto the plot figure.
In my imlpementation, when a checkbox changes state, it can send two distinct signals.
One of the signals appends the desired value to the temp dict and the other removes from the dict.
This way, the user fully controls what changes are made from the two checkboxes provided.